\docType{data}
\name{aphids}
\alias{aphids}
\title{Aphids data set}
\value{
  A data frame
}
\description{
  The is data described in Matis et al, 2008. The data set
  consists of five observations on cotton aphid counts on
  twenty randomly chosen leaves in each plot, for
  twenty-seven treatment-block combinations. The data were
  recorded in July 2004 in Lamesa, Texas. The treatments
  consisted of three nitrogen levels (blanket, variable and
  zero), three irrigation levels (low, medium and high) and
  three blocks, each being a distinct area. Irrigation
  treatments were randomly assigned within each block as
  whole plots. Nitrogen treatments were randomly assigned
  within each whole block as split plots. . Note that the
  sampling times are $t$=0, 1.14, 2.29, 3.57 and 4.57 weeks
  (i.e. every 7 to 8 days). \describe{ \item{Time}{Sampling
  time (in weeks). This has been slightly simplified}
  \item{Water}{Water level at that particular plot: Low,
  Medium and High} \item{Nitrogen}{Nitrogen level at that
  plot: Blanket, Variable and Zero} \item{Block}{The plot
  block: 1, 2 or 3} \item{Aphids}{The number of aphids
  counted} }
}
\keyword{datasets}

