\docType{data}
\name{Beauty}
\alias{Beauty}
\title{Beauty data set}
\value{
  A data frame
}
\description{
  This data set is from a study where researchers were
  interested in whether a lecturers' attractiveness
  affected their course evaluation. This is a cleaned
  version of the data set and contains the following
  variables: \describe{ \item{evaluation}{the questionnaire
  result} \item{tenured}{does the lecturer have tenure; 1
  == Yes. In R, this value is continuous}
  \item{minority}{does the lecturer come from an ethnic
  minority (in the USA)} \item{age}{the lecturers' age}
  \item{gender}{a factor: Female or Male}
  \item{students}{number of students in the class}
  \item{beauty}{each of the lecturers' pictures was rated
  by six undergraduate students: three women and three men.
  The raters were told to use a 10 (highest) to 1 rating
  scale, to concentrate on the physiognomy of the professor
  in the picture, to make their ratings independent of age,
  and to keep 5 in mind as an average. The scores were then
  normalised.} }
}
\keyword{datasets}

