\docType{data}
\name{cell_data}
\alias{cell_data}
\title{Dummy cell data set}
\value{
  A data frame
}
\description{
  Example cell data set. An experiment was conducted
  involving two cell types (Case and Control) and two
  treatments (A and B). The data is stored as a data frame:
  \describe{ \item{values}{measurements from the
  experiment} \item{treatment}{either A or B}
  \item{type}{Case or Control} }
}
\keyword{datasets}

