\docType{data}
\name{google}
\alias{google}
\title{Google data set}
\value{
  A data frame
}
\description{
  The google data set \describe{ \item{Rank}{Site rank (in
  terms of users)} \item{Site}{Site name}
  \item{Category}{Site classification}
  \item{Users}{Approximate number of users}
  \item{Views}{Approximate page views}
  \item{Advertising}{Does the site have advertising} }
}
\keyword{datasets}

