% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nclRggplot2-package.R
\docType{package}
\name{nclRggplot2-package}
\alias{nclRggplot2}
\alias{nclRggplot2-package}
\title{R courses at Newcastle University}
\description{
Functions and datasets used in for external R courses at Newcastle University
}

